requirements.txt
pycryptodome>=3.20.0
protobuf>=4.25.0
pyserial>=3.5
